Use a ternary condition to do conditional rendering
The ternary operator is often used as a replacement for if/else statements in Javascript, in React they are also very popular.

One reason for this is because of how JSX is compiled, if/else statements cannot be inserted directly into JSX code. When you need to use the if/else, they were always left out of the return statement, as with the ternary expressions they can be used within the JSX element, which can be an alternative if you want to create conditional logic within your JSX element.

The ternary operator has three mandatory parts, but you can also do other ternary combinations. The syntax of the ternary operator is like this:

condition? run if true : else run this

In the code, we have three constant variables called buttonOne, buttonTwo and buttonThree defined in the render() method of the CheckUserAge component, each of them receives a JSX element that represents a button element.

Initialize component state CheckUserAge with input and userAge set to an empty string. In the component's return statement, use the ternary operator by implementing the following logic:

when the page first loads, render the submit button, buttonOne, for the page. Then, when the user enters their age in the field and clicks the button, a button will be rendered with a message according to the age that was entered for the user.

If the user enters a number greater than or equal to 18 and clicks the button, the button's inner text will change to You May Enter on that user's screen. If the user enters a smaller number the inner text of the button will change to the text You Shall Not Pass.