Conditionally render elements from Component props
An interesting topic in React is that we can render elements from a prop of a component, using props to conditionally render code is very common among developers using React, they use the value and a certain prop to determine what will be rendered.

We are creating a game where the user presses the button and he will see if he won or not. To create this game we used two components, a parent called GameOfChance and a child called Results.

The child component Results we will configure it so that it will be responsible for making the decisions of what will be rendered based on the value of the prop that is passed to it and the parent component GameOfChance will be responsible for the visual part of the game with the click button, counter and rendering the result on the screen.

In the constant expression variable we will store an expression that will return a different value every time it is executed, so we will use Math.random(). This method returns a value between 0 (inclusive) and 1 (exclusive) each time it is called.

So in the constant expression we will store this Math.random() >= .5 expression, this means that this expression will return true 50% of the time and false the other 50%.

Render the Results child component as a child of GameOfChange and pass the expression to a prop called fiftyFifty.

Now in the Results component, write a ternary expression to render the h1 element with the text You won! or you lose! according to the value of the fiftyFifty prop being passed to Results. Increment the counter value by 1 to make sure the counter is counting correctly.