Change input border color conditionally based on state of React component
With React you can also conditionally render a CSS style based on the state of the component. For example you can change a CSS style when the user writes more than 10 characters or send a message when he sends his data less than 10 characters and so on.

It is very important to understand how this works because in this way it is very efficient to manipulate the states of our components and we can follow when elements change and how we will deal with them when they change, unlike you applying and defining new styles by modifying DOM elements directly (which is very common with jQuery).

In the code, we have a component called GateKeeper that has input field controlled with a styled border with 1px solid black and we want that if the user types more than 15 characters in the input field, set the border style to 3px solid red.

Inside the component's render method, create a condition if the user types more than 15 characters in the field, change the input border to 3px solid red.