Use array method map() to render a list of items
Conditionally rendering is useful, but you may need your components to render an unknown number of elements. For example, you have your application which is a shopping list where the user will be putting on that list what he needs to buy in the market.

As a programmer, you don't know how many items the user will put in the list, so you need to configure your component so that every time the user adds a new item and clicks the add button, the correct number of items in the list is dynamically rendered. list.

We have a component called MyToDoList that has the visual part of the screen already defined, we just have to make sure that our To-Do List application is working correctly. Inside the constructor, create a this.state object and define two states: the userInput will be initialized as an empty string and the toDoList must be initialized as an empty array.

Then delete the null value of the items constant variable that is inside the render() method, instead we will use the map method to map the toDoList array that is stored inside the state and we will dynamically render a li for each item.

So when the user writes a to-do in the field and clicks the add button, he will have put that to-do in the list.