Give sibling li elements a unique key attribute
When we create an array of elements, each element in that array needs to have a unique key, React uses this to keep track of which items are added, changed, or removed. This helps make the rendering process more efficient when the list changes in any way. Keys only need to be unique between sibling elements like li tags for example.

In the code we have a constant variable called frontEndFrameworks that stores an array with some front-end frameworks and a functional stateless component called Frameworks( ). Frameworks( ) will map the array to an unordered list and return a li element with the key attribute for each li, the li elements will have the text of each front-end framework in the array.