Use array filter method to filter online users
The filter( ) method receives a callback function as a parameter, the filter will filter all the contents of an array and return a new array with the elements passed in the callback function.

In the code, we have a component MyComponent that in the state of this component there is an array of users, with some users online and others not. Filter the array of users and return a new array containing only those users whose online property is true. Store this operation in the usersOnline variable.

Then in the renderOnline variable, map the filtered array and return a li element with the name of that user, each li will have a unique key, for that, add the key attribute to the li element.