Render React on server using renderToString method
So far, we've rendered React components on the client. Normally, this is what you will always do. However, there will be some cases where we need to render a component on the server. As React is a Javascript visualization library, it is allowed to run Javascript code on the server with Node, React provides a method called renderToString() you pass a React element, React will return an HTML string.

With this we improve the performance of our application because we will only depend on the resources of the server and the speed of the user's network, this is very useful for sites that have a lot of content.

The renderToString() method is provided in ReactDOMServer, which is available here as a global object, use the method to render the App component to a string.