Create a Redux store
Redux is a library that manages the state of Javascript applications. Redux came to fill this need to control the states in each component as we do in React. Taking the responsibility of each of the components to store the states, leaving everything centralized, being used at the same time by all the components in a shared way.

In Redux, you will have a single place that will be responsible for the entire state of the application. This means that if you had a React app with ten components and each component having its own state locally, you would put all of your app's state into a single state object, that single place is called the store.

The store will be the place where all the states of our application will be. So whenever any part of your app's state updates, it will call the pure function called reducer that takes the state it wants to update and the action that represents what will be put into that new state and returns that new state.

In the code, we have the reducer defined , create a store variable and assign the createStore( ) method to it, passing the reducer function as an argument.