Get the state of the Redux store with the getState() method
The getState() method returns the current state of the store and if we want to update the state of our application we will need this method.

In the code, we already have the store already defined and now we need to get its current state, use store.getState() to get its current state and assign it to a variable called currentState.