Define a Redux action object
Since Redux is a state management framework, updating state is one of its main tasks. In Redux, all state updates that are made are through the action object.

The action object is an action that contains information about an action event that has occurred. Then, the Redux store receives this action object and then updates its state with that action that happened. For example, the action loads a username after a user logs in. Although optional, actions objects must have a type property that specifies the 'type' of action that took place.

Actions provide information about events that occurred in your application to the Redux store. The store conducts business and updates the state based on the action that took place.

Create an action object and give it the type property with the string value 'LOGIN'.