Define an action creator function to send an action to the store
After we created an action, now we have to send this action to the Redux store so that it updates its state.

In Redux, you define action creators to perform this process. The action builder is a Javascript function that returns an action. Then it creates objects that represent actions that occur in our application.

Define a function called actionCreator() that returns the action object when called.