Dispatch an event to the Redux store with the dispatch( ) method
The dispatch method is what you use to dispatch the actions that occurred to the Redux store, for that we use this syntax: store.dispatch( ) as a parameter passing the value returned from an action creator which is the action we want to send to the store .

In the code, we have the Redux store with an initialized state which is an object containing a login property currently set to false. There is also an action builder called loginAction that returns an action of type LOGIN.

Dispatch the action of type LOGIN to the Redux store by calling the dispatch method, passing the loginAction( ) creator as a parameter to the method.