Manage an action in the store using a reducer function
After an action is created and dispatched, the Redux store needs to know how it will respond to that action. For this, Redux provides us with reducing functions, they are what do this job. The reducers are responsible for handling these actions that occur in our application, so it receives this information, handles it and sends it to the store.

A reducer is a pure Javascript function that takes the previous state or the state defined with pattern and an action as arguments, and returns a new state.

In the code, we already have a reducer function already defined, inside the reducer function we will check the action type, if the action type is 'LOGIN', we will return a state object with the login property being set to true. Otherwise, we will return the current state.

To check the action type we use this syntax: action.type.