Use a switch statement to handle multiple types of actions
You can tell the Redux store how to handle various action types. Let's say you are managing user authentication in your Redux store. You want to have a state representation for when users are logged in and when they are logged out. You can do this with a single state object with an authenticated property, we will also need action creators to represent the logged in user and the logged out user.

In the code we have a store, actions and action creators already defined, now we need to deal with various types of authentication that can occur. Use a Javascript switch statement inside the reducer function to handle and respond to these different authentication actions.

The switch statement will receive the action type, if the action type is 'LOGIN', we will return an object with the authenticated property set to true. If the type is 'LOGOUT', we will return an object with the authenticated property set to false, if it is none of these cases, we will return the current state.