Use const to store user action types
A good practice when working with Redux is to assign action types to a constant variable to make our code more organized and easier to read, and then we can reference these constants elsewhere if we need to.

Declare a const called LOGIN that will store the string 'LOGIN' and another const called LOGOUT that will store the string 'LOGOUT'.

Next, edit the authReducer() reducer function and action creators so that they reference LOGIN and LOGOUT constants instead of string values.