Notify a store change using the store.subscribe() method
The store.subscribe() method registers a listener function for the Redux store, every time an action is dispatched to the store, this subscribe function will be executed.

Write a callback function that increments the global variable count every time the store receives an action, and pass that function to the store.subscribe() method.

store.dispatch() is called three times in a row, each time passing an action object with the ADD type property. So when the ADD action is dispatched in the store, the counter must be incremented by 1.