Combine multiple reducing functions in the combineReducers() method
The combineReducers() method allows the combination of several reducer functions within an object with the keys associated with that specific reducer function. The name of the keys we define will be used by Redux to store the state of that specific reducer function.

We can define multiple reducers to handle different parts of our app's state and then compose those reducers together into a root reducer, hence the root reducer is passed to Redux's createStore() method.

In the code editor we already have two reducer functions called counterReducer() and authReducer(), countReducer() is responsible for the counter of our application, authReducer() is responsible for taking care of authenticated and unauthenticated users.

In the rootReducer() function store the combination of the counterReducer() and authReducer() reducer functions in an object defining count and auth properties. Assign counterReducer to the count key and authReducer to the auth key.