Send the action data to the store
Redux allows you to send specific data of an action along with that action, actions usually originate from some interaction that the user made and they tend to carry some data along with these actions and the Redux store needs to know this information about the action that occurred.

In the code editor we already have the type of action created at the top of the code, a reducer function notesReducer() and an action creator addNoteText() that receives a note. Complete the body of the addNoteText() function so that it returns an action object with the type of that action and data of that action which is the text of the note. When calling the action creator, we'll pass specific note information that can be accessed to the object.

Then finish writing the switch statement in notesReducer(). Write a case that if the action type is ADD_NOTE we will return the text property of the received action as the new state.

When the action is dispatched, the state of the store that was the value of the string Initial State will be updated with the value of the text of the note the string 'Hello!'.