Use middleware to handle asynchronous actions in the application
Redux provides us with a tool that allows us to include custom functions that process the action that occurred and deal with this action before dispatching the action and sending it to the reduce, this tool is called the Redux Thunk middleware. The middleware that handles the types of requests we receive in our application.

To include the Redux Thunk middleware, we pass it as an argument to Redux.applyMiddleware(). This statement is then provided as an optional second parameter to the createStore() function. Then, to create an asynchronous action, you return a function in the action's creator that takes dispatch as an argument. In this role, you can dispatch actions and perform asynchronous requests.

In the code editor, we have an example of a simulated asynchronous request with a setTimeout() call. It's common to dispatch an action before initiating any asynchronous behavior so that your app's state knows that some data is being requested (this state might display a loading icon, for example). Then, after receiving the data, you dispatch another action that loads the data as a payload along with the information that the action has completed.

Write two dispatches in the handleAsync action builder, the first one dispatches the requestingData() asynchronous action before the setTimeout(). After we receive this (imaginary) data. dispatch the ReceiveData() asynchronous action, passing that data.