Create a counter with Redux
Create a counter that will have 0 as the initial state, at the top of the code editor, we will have the action types INCREMENT and DECREMENT.

Below we will have the reducing function called counterReducer that receives the initial state and the action, inside the function we will check what type of action occurred with the switch statement, if the type of this action is INCREMENT, we will increment the counter state by 1. If it is a decrement, we will decrement the counter state by minus 1, if it is none of these cases, we will return the current state.

We are going to use two action creators called incAtion() and decAction, the incAction will return an action object of type INCREMENT, while the decAction will return an action object of type DECREMENT.

When finished, we can dispatch INCREMENT or DECREMENT actions to increment or decrement the state held in the store.