Return a new copy of the state instead of changing it directly
This principle is called immutability, an immutable state means that you never modify the state directly, instead you return a new copy of the state.

Redux does not actively enforce state immutability on its storage or reducers, that responsibility lies with the programmer. Fortunately, JavaScript (especially ES6) provides several useful tools that you can use to enforce the immutability of your state, be it a string, number, array, or object. Note that strings and numbers are primitive values and are immutable in nature. In other words, 3 is always 3. You cannot change the value of the number 3. An array or object, however, is mutable. In practice, your state will likely consist of an array or object, as these are useful data structures for representing many types of information.

In the code editor, we see a store created, the action type at the top of our code, a reducer function immutableReducer that takes a list of tasks called todos and the action, and an action creator called addToDo that returns an action object with the data for that action.

In the switch statement, we check the action type, if this action is of the ADD_TO_DO type, we will return a new array with the action.todo item, if the action is not of this type, we will return the current state.