Use the Spread Operator to pass current state and action values to a new array
The spread operator has this syntax here when you want to pass values from an array to a new array: [...values]. It's an ES6 solution that can help us with this issue of state immutability in Redux.

Use the spread operator to return a new one with the current state values and the new task.