Remove an item from an array and return a copy of the state with the slice() method
In the code editor, we have an action creator called removeItem that receives a specific index of an item and returns an action object with the type of action and the index of the item.

In the reducer function immutableReducer we receive an initial state with an array containing six numbers and an action object, inside the reducer function we check the type of that action. If this type is REMOVE_ITEM, we will remove an item from the state and return a copy of that state array without that item in the specific index of that array, if the action is not of that type, we return the current state.

So, if the action is of the remove item type, we return a copy of the state without changing the state directly and with this copy without that item of that specific index in it.