Return a copy of an object using the Object.assign() method
The principle of immutability of the state is not only done with arrays, it is also used when the state is an object. A useful tool for dealing with objects is the Object.assign() utility. The Object.assign() method copies only its own and enumerable properties from a source object to a target object.

Object.assign() receives as its first parameter an empty object that represents where we will place these properties followed by the object(s) we wish to copy.

In the code editor specifically within the reducing function, make sure that if the action type is ONLINE, it returns a new state object that has the status property defined with the string online.