Introduction of React and Redux
For us to use these two libraries together, we have to know beforehand how they work and how we can use them in a web application.

React is a Javascript library used for creating user interfaces, in it we separate an application into small parts that we call components, each component of the page having a responsibility on its page, each component has its state created within the component. Redux is a state management structure that stores all the state of an application in one place, this place is called the store.

In a React Redux app, you create the store that will manage the state of your entire app, so your components can directly access the data in the store and get the necessary state that component needs.

We'll create a React component that allows the user to enter new text messages into the input, these text messages will be added to an array that will be displayed in the view.

In the code editor we have already defined the class component called DisplayMessages, inside the component, add a constructor and the super keyword passing props to both. Then initialize it with a state containing two properties: input set to an empty string and messages set to an empty array.

The DisplayMessages component must render an empty div element.