Manage component state locally
We will finish creating our DisplayMessages class component. We have already defined inside the class component, the state with the properties input and messages initialized with an empty string and an empty array.

Now to finish our application, in the render() method, we will make the component render an input element, a button element and a ul element. When the input element's value changes, it should add a handleChange() method. When the button is clicked, it should add the submitMessage() method.

Then write handleChange() and submitMessage() methods. The method should update the input with whatever value the user is typing, the submitMessage() method should concatenate the current message which is what the user wrote stored in the input to the local state message array and after the user wrote the message and clicked the button, we will make the input field empty again.

And finally, we'll use the ul element to map this array of messages and render the array's messages to the screen as a list of li elements.