Use the Redux store for state management
Now we will use the Redux store to manage the states of our application, we made a React component that has its own state and makes all the logic for the component work, now we need to move all this logic from this component to Redux.

The only functionality the app has is adding new user messages to an unordered list. Now we'll see how React and Redux work together.

First define an action type ADD and set it to ADD const. Next, define an addMessage() action builder that builds the action to add a message. We'll need to pass a message to this action builder and include the message in the returned action.

Then create a reducer function called messageReducer() that will handle the state of the messages, the initial state must be an empty array. This reducer must either add a message to the array of messages held in state or return the current state. Finally, create the Redux store and pass the reducer.