Use provider to connect Redux with React
Now that we've already created a Redux store to handle the message array and a creator action to add new messages. We will continue to develop our application. The next step is for us to provide React access to the Redux store and the actions needed to dispatch updates.

React Redux provides an API that makes React child components available to access the state of the Redux store, this API is called Provider. The Provider uses two supports, the Redux store and the application's child components. An example of a provider definition for an application component:

<Provider store={store}>
   <App/>
</provider>

At the end of the code editor, we have a component called AppWrapper, we will use this component to render the React Redux provider and pass the store store as props, it should be the same as the Redux store. Then render the DisplayMessages component as a child.

After all this process is done, we will see the React component rendered on the page.