Map store state and send it to a React component prop
The Provider component allows you to provide state and submission for your React components, but you must specify exactly what state and actions you want. That way, you ensure that each component only has access to the state it needs. To pass the state of the store to the React component, we'll use a function called mapStateToProps().

This function will receive the state of the store that is defined with an empty array as an argument and it will return an object with a messages property defined with the value of the state.

This property may be accessible to a component through props.