Dispatch action builders as props to React components
We will use the mapDispatchToProps() function, this function is used to provide specific action creators to your React components so they can dispatch actions against the Redux store.

It is an anonymous function that receives a dispatch parameter as a parameter and returns an object and maps the dispatch actions to property names and passes them to the component as properties ( props ).

In the code editor, we have an action creator called addMessage, write the function mapDispatchToProps() receiving dispatch as an argument and returning an object. The object must have a submitNewMessage property defined for the dispatch function, which takes a parameter for the new message to be added when dispatching addMessage().