Connect Redux with React via connect method
Now that we've written the mapStateToProps() and mapDispatchToProps() functions, we can use them to map the state and dispatch to the props of one of your React components. We'll use the connect method to connect a React component to Redux.

The connect method receives the functions mapStateToProps() and mapDispatchToProps() as arguments and calls a component.

The code editor has mapStateToProps() and mapDispatchToProps() functions and a new React component called Presentational. Connect the Presentational component to Redux with the connect method and call it immediately on the Presentational component. Assigns the result to a new const called ConnectedComponent representing the connected component.