Connect Redux to Messaging App
Now we are going to connect Redux to the React component that handles messages that we did in some previous challenges.

In the last lesson of this module, the component we hooked up to Redux was named Presentational, and that wasn't arbitrary. This term usually refers to React components that are not directly connected to Redux. They are simply responsible for the presentation of the UI and they do this based on the props they receive. On the other hand, container components are connected to Redux. Typically they are responsible for dispatching actions to the store and often pass the state of the store to child components as props.

In the code editor we have the React component that has been renamed to Presentational. Create a new component held in a constant called Container that uses connect to connect the Presentational component to Redux.

Then, in AppWrapper, render the React Redux Provider component. Pass the Provider to the Redux store as a prop and render the Container as a child. Once everything is set up we will see the messaging app rendered on the page again.