Extract local state of Presentational component to Redux
We've already prepared the Redux code so that it could control the state management of the React messaging app. Now that Redux is connected, we need to extract the state management from the Presentational component and place it in Redux. We currently have Redux connected, but the component is still handling state locally.

The first thing we're going to do inside the Presentational component is: remove the messages property from the component's local state. These messages will be managed by Redux. Then modify the submitMessage() method so that it sends submitNewMessage() from this.props and passes the current local state message entry as an argument. Since we removed the messages from the local state, remove the messages property from the call to this.setState() here as well. Finally, modify the render() method so that it maps messages received from props instead of state.

After these changes are made, the application will continue to work in the same way, only now Redux is managing the state.