Introduction to when to use Redux and React on your machine locally
In the challenges we made from this module here in Redux together with React, we used the freecodecamp code editor, but when we use Redux and React in an application we will have to use a local machine.

Probably using npm to help you with your website development and also you can use Babel to add React and ReactDOM as external JavaScript resources and work there too.

In the code editor, we have inside a comment basically what you would use to develop your application on your local machine. Below the comment, enter the string 'Now I know React and Redux!' into the console.